tests and confirmation msg

+++ src/commands/utils/deleteArtifact.ts
@@ -1,16 +1,24 @@
 import * as vscode from "vscode";
 import { DeleteArtifactV1FlinkArtifactRequest } from "../../clients/flinkArtifacts/apis/FlinkArtifactsArtifactV1Api";
+import { artifactUploadDeleted } from "../../emitters";
+import { FlinkArtifact } from "../../models/flinkArtifact";
 import { EnvironmentId, IEnvProviderRegion } from "../../models/resource";
 import { getSidecar } from "../../sidecar";
 
 export async function deleteArtifactCommand(
-  params: DeleteArtifactV1FlinkArtifactRequest,
+  selectedArtifact: FlinkArtifact | undefined,
 ): Promise<void> {
+  if (!selectedArtifact) {
+    void vscode.window.showErrorMessage(
+      "Cannot delete artifact: missing required artifact properties.",
+    );
+    return;
+  }
   const request: DeleteArtifactV1FlinkArtifactRequest = {
-    cloud: params.cloud,
-    region: params.region,
-    environment: params.environment,
-    id: params.id,
+    cloud: selectedArtifact.provider,
+    region: selectedArtifact.region,
+    environment: selectedArtifact.environmentId,
+    id: selectedArtifact.id,
   };
   const sidecarHandle = await getSidecar();
   const providerRegion: IEnvProviderRegion = {
@@ -20,9 +28,24 @@ export async function deleteArtifactCommand(
   };
   const artifactsClient = sidecarHandle.getFlinkArtifactsApi(providerRegion);
 
-  const response = await artifactsClient.deleteArtifactV1FlinkArtifact(request);
-  // what type should response be?
+  const yesButton = "Yes, delete";
+  const confirmation = await vscode.window.showWarningMessage(
+    `Are you sure you want to delete "${selectedArtifact.name}"?`,
+    {
+      modal: true,
+      detail:
+        "Deleting this artifact will disable all User-Defined Functions (UDFs) created from it. Consequently, any Flink statements that utilize these UDFs will also fail. This action cannot be undone.",
+    },
+    { title: yesButton },
+    // "Cancel" is added by default
+  );
+  if (confirmation?.title !== yesButton) {
+    return;
+  }
+
+  await artifactsClient.deleteArtifactV1FlinkArtifact(request);
+  artifactUploadDeleted.fire();
   void vscode.window.showInformationMessage(
-    `Artifact "${response}" deleted successfully from Confluent Cloud.`,
+    `Artifact "${selectedArtifact.name}" deleted successfully from Confluent Cloud.`,
   );
 }
